test: migrate 3 small route tests to bdd (round 42)
Migrates 3 small legacy route tests to BDD format:

- integrations-github-delete (8→3 BDD it()s)
- test-telegram-mock (9→2 BDD it()s)
- model-stats (4→2 BDD it()s)

21 legacy it()s → 7 BDD it()s (67% reduction). No per-file
coverage regression. api.bdd.md updated with Round 42 entry.

The model-stats.bdd.test.ts adds an afterEach cleanup that
wipes modelStat and modelUsageObservation rows to avoid
stale data from prior runs (the legacy test had no
cleanup and was always potentially flaky when rerun).
